rename files and add custom colors to caves
